backend/requirements.txt
# ============================================
# InvoiceGhost Backend Dependencies
# Privacy-first invoice & receipt parser SaaS
# ============================================

# ============================================
# Core Framework
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ============================================
# Rate Limiting
# ============================================
slowapi==0.1.9

# ============================================
# PDF & Image Processing
# ============================================
pdf2image==1.16.3
Pillow==10.2.0

# ============================================
# AI Integration
# ============================================
google-generativeai==0.3.2
groq==0.4.1

# ============================================
# Development & Testing
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# ============================================
# Code Quality & Type Checking
# ============================================
ruff==0.1.9
mypy==1.8.0

# ============================================
# Security & Utilities
# ============================================
cryptography==41.0.7  # For HMAC license key validation
